requirements.txt
asgiref==3.5.2
Django==3.2
django-cors-headers==3.10.1
django-crispy-forms==1.14.0
django-filter==21.1
django-widget-tweaks==1.4.12
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
gunicorn==20.1.0
Pillow==9.1.1
PyJWT==2.4.0
pytz==2022.1
sqlparse==0.4.2
whitenoise==5.2.0
